+++ examples/scripts/conditional.py
@@ -30,6 +30,7 @@ app('cd {}'.format(directory))
 if self._last_result:
     print('\nContents of directory {!r}:'.format(directory))
     app('dir -l')
+    print('{}\n'.format(self._last_result.data))
 
     # Change back to where we were
     print('Changing back to original directory: {!r}'.format(original_dir))